# internal/routes/api/V1/tag.go
package V1

import (
	"blog_service/internal/service"
	"blog_service/pkg/app"
	"blog_service/pkg/convert"
	"blog_service/pkg/errcode"
	"github.com/gin-gonic/gin"
)

type Tag struct {
}

func (t *Tag) Get(c *gin.Context) {

}

// @Summary 获取多个标签
// @Produce json
// @Param name body string true "标签名称" minlength(3) maxlength(100)
// @Param state query int false "状态" Enums(0,1) default(1)
// @Param page query int false "页码"
// @Param page_size query int false "每页数量"
// @Success 200 { object } model.TagSwagger  <-- This is a user defined struct "成功"
// @Failure 400 { object } errcode.Error  <-- This is a user defined struct "请求错误"
// @Failure 500 { object } errcode.Error  <-- This is a user defined struct "内部错误"
// @Router /api/v1/tags [get]
func (t *Tag) List(c *gin.Context) {
	param := service.TagListRequest{}
	response := app.NewResponse(c)
	_, errs := app.BindAndValid(c, &param)
	if errs != nil {
		response.ToErrorResponse(errcode.InvalidParams.WithDetails(errs.Errors()...))
		return
	}
	svc := service.NewTagService(c.Request.Context())
	pager := app.Pager{
		Page:     app.GetPage(c),
		PageSize: app.GetPageSize(c),
	}
	totalRows, err := svc.CountTag(&service.CountTagRequest{Name: param.Name, State: param.State})
	if err != nil {
		response.ToErrorResponse(errcode.ServerError)
		return
	}
	tags, err := svc.GetTagList(&param, &pager)
	if err != nil {
		response.ToErrorResponse(errcode.ServerError)
		return
	}
	response.ToResponseList(tags, totalRows)
	return
}

// @Summary 新增标签
// @Produce json
// @Param name body string true "标签名称" minlength(3) maxlength(10)
// @Param state body int false "状态" Enums(0,1) default(1)
// @Param created_by body string false "创建者" minlength(3) maxlength(10)
// @Success 200 { object } model.TagSwagger <-- This is a user defined struct "成功"
// @Failure 400 { object } errcode.Error <-- This is a user defined struct "请求错误"
// @Failure 500 { object } errcode.Error <-- This is a user defined struct "内部错误"
// @Router /api/v1/tags [post]
func (t *Tag) Create(c *gin.Context) {
	param := service.CreateTagRequest{}
	response := app.NewResponse(c)
	_, errs := app.BindAndValid(c, &param)
	if errs != nil {
		response.ToErrorResponse(errcode.InvalidParams.WithDetails(errs.Errors()...))
		return
	}
	svc := service.NewTagService(c.Request.Context())
	if err := svc.CreateTag(&param); err != nil {
		response.ToErrorResponse(errcode.ServerError.WithDetails(err.Error()))
		return
	}
	response.ToResponse(gin.H{})
}

// @Summary 更新标签
// @Produce json
// @Param id path int true "标签ID"
// @Param name body string true "标签名称" minlength(3) maxlength(10)
// @Param state body int false "状态" Enums(0,1) default(1)
// @Param modified_by body string false "修改者" minlength(3) maxlength(10)
// @Success 200 { array } model.Tag "成功"
// @Failure 400 { object } errcode.Error <-- This is a user defined struct "请求错误"
// @Failure 500 { object } errcode.Error <-- This is a user defined struct "内部错误"
// @Router /api/v1/tags/{id} [put]
func (t *Tag) Update(c *gin.Context) {
	id := convert.StrTo(c.Param("id")).MustUInt32()
	param := service.UpdateTagRequest{ID: id}
	response := app.NewResponse(c)
	_, errs := app.BindAndValid(c, &param)
	if errs != nil {
		response.ToErrorResponse(errcode.InvalidParams.WithDetails(errs.Errors()...))
		return
	}
	svc := service.NewTagService(c.Request.Context())
	if err := svc.UpdateTag(&param); err != nil {
		response.ToErrorResponse(errcode.ServerError.WithDetails(err.Error()))
		return
	}
	response.ToResponse(gin.H{})
	
}

// @Summary 删除标签
// @Produce json
// @Param id path int true "标签ID"
// @Success 200 { string } string "成功"
// @Failure 400 { object } errcode.Error "请求错误"
// @Failure 500 { object } errcode.Error "内部错误"
// @Router /api/v1/tags/{id} [delete]
func (t *Tag) Delete(c *gin.Context) {
	id := convert.StrTo(c.Param("id")).MustUInt32()
	param := service.DeleteTagRequest{ID: id}
	response := app.NewResponse(c)
	if _, errs := app.BindAndValid(c, &param); errs != nil {
		response.ToErrorResponse(errcode.InvalidParams.WithDetails(errs.Errors()...))
		return
	}
	svc := service.NewTagService(c.Request.Context())
	if err := svc.DeleteTag(&param); err != nil {
		response.ToErrorResponse(errcode.ServerError.WithDetails(err.Error()))
		return
	}
	response.ToResponse(gin.H{})
	
}

func (t *Tag) ChangeState(c *gin.Context) {

}
func NewTag() Tag {
	return Tag{}
}